API/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-aiplatform
vertexai
python-multipart
requests
python-dotenv
xmltodict
httpx
